Split local git tool tests
