+++ docs/DECISIONS.md
@@ -331,3 +331,4 @@
 - Terminal run lifecycle status planning lives in `WorkspaceTerminalLifecyclePlanner`. `WorkspaceModel.runTerminalCommand` still owns async streaming, terminal entry mutation, session marker cleanup, and top-bar application, but started, missing-context, stopped, cancelled, and final idle/failed status selection should stay in the focused planner so terminal behavior stays aligned with tool-run lifecycle boundaries.
 - Active-work Stop All and Disconnect All lifecycle status planning lives in `WorkspaceActiveWorkStopPlanner`. `WorkspaceModel` still owns cancelling MCP servers, send state, terminal state, remote-project detachment, and top-bar application, but cancel/disconnect error clearing plus stopped/idle status selection should stay in the focused planner so toolbar commands follow the same testable lifecycle boundary as tool and terminal runs.
 - Command routing must be explicit. `WorkspaceCommandRoutingCatalog` classifies host-owned commands such as Stop All and Computer Use permission actions separately from command IDs that `WorkspaceCommandPlan` can execute in the workspace model. SwiftUI and desktop planners reject unknown IDs instead of dispatching them into silent no-ops, and command-surface coverage proves every emitted command is either presentational or dispatchable.
+- Local git, git hunk, git worktree, and git router coverage lives in focused QuillCodeTools suites, not the mixed `ToolTests.swift` catch-all. `GitLocalToolExecutorTests`, `GitPatchToolExecutorTests`, `GitWorktreeToolExecutorTests`, and `GitToolRouterTests` share fixtures from `ToolTestSupport.swift` so git behavior can evolve without turning one broad file into a failure triage bottleneck.
